Conditionally run activity destroy hooks
In case when app is launched directly via NavigationAcitivtyin onCreate we shut that activity down immediately and start SplashActivity. This causes an issue when the context gets initialized in splash but then immediately killed by the NavigationActivity getting destroyed.

Here we add a new flag `resettingContext` that is set whenever we shut down NavActivity in order to launch SplashActivity. In that case since we skip the whole initialization on the activity we can also skip all the pause/stop/destroy hooks